requirements.txt
Django==4.2.10
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
celery==5.5.1
redis==5.2.1
python-dotenv==1.1.0
pytest==8.3.5
pytest-django==4.11.1
google-cloud-language==2.17.1
google-cloud-translate==3.20.2
google-cloud-aiplatform==1.89.0
djangorestframework-simplejwt==5.5.0
Pillow==11.2.1
